+++ android/src/test/kotlin/com/shareinvest/flutter_device_platform_id/FlutterDevicePlatformIdPluginTest.kt
@@ -1,14 +1,14 @@
-package com.shareinvest.flutter_device_unique_id
+package com.shareinvest.flutter_device_platform_id
 
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlin.test.Test
 import org.mockito.Mockito
 
-internal class FlutterDeviceUniqueIdPluginTest {
+internal class FlutterDevicePlatformIdPluginTest {
     @Test
     fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
-        val plugin = FlutterDeviceUniqueIdPlugin()
+        val plugin = FlutterDevicePlatformIdPlugin()
 
         val call = MethodCall("getPlatformVersion", null)
         val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)