+++ example/macos/RunnerTests/RunnerTests.swift
@@ -5,7 +5,7 @@ import XCTest
 // If your plugin has been explicitly set to "type: .dynamic" in the Package.swift,
 // you will need to add your plugin as a dependency of RunnerTests within Xcode.
 
-@testable import flutter_device_unique_id
+@testable import flutter_device_platform_id
 
 // This demonstrates a simple unit test of the Swift portion of this plugin's implementation.
 //
@@ -14,7 +14,7 @@ import XCTest
 class RunnerTests: XCTestCase {
 
   func testGetPlatformVersion() {
-    let plugin = FlutterDeviceUniqueIdPlugin()
+    let plugin = FlutterDevicePlatformIdPlugin()
 
     let call = FlutterMethodCall(methodName: "getPlatformVersion", arguments: [])