+++ windows/include/flutter_device_platform_id/flutter_device_platform_id_plugin_c_api.h
@@ -1,5 +1,5 @@
-#ifndef FLUTTER_PLUGIN_FLUTTER_DEVICE_UNIQUE_ID_PLUGIN_C_API_H_
-#define FLUTTER_PLUGIN_FLUTTER_DEVICE_UNIQUE_ID_PLUGIN_C_API_H_
+#ifndef FLUTTER_PLUGIN_FLUTTER_DEVICE_PLATFORM_ID_PLUGIN_C_API_H_
+#define FLUTTER_PLUGIN_FLUTTER_DEVICE_PLATFORM_ID_PLUGIN_C_API_H_
 
 #include <flutter_plugin_registrar.h>
 
@@ -13,11 +13,11 @@
 extern "C" {
 #endif
 
-FLUTTER_PLUGIN_EXPORT void FlutterDeviceUniqueIdPluginCApiRegisterWithRegistrar(
+FLUTTER_PLUGIN_EXPORT void FlutterDevicePlatformIdPluginCApiRegisterWithRegistrar(
     FlutterDesktopPluginRegistrarRef registrar);
 
 #if defined(__cplusplus)
 }  // extern "C"
 #endif
 
-#endif  // FLUTTER_PLUGIN_FLUTTER_DEVICE_UNIQUE_ID_PLUGIN_C_API_H_
+#endif  // FLUTTER_PLUGIN_FLUTTER_DEVICE_PLATFORM_ID_PLUGIN_C_API_H_